+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to the "mcp--inspector" extension will be documented in this
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.0.1] - 2025-03-29
 
 - Initial release
+- Tool List Support
+- Transport Support (STDIO / SSE)
+
+## [0.0.2] - 2025-03-29
+
+- Tool Execution and Viewer Support
+